requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
aiohttp=3.7.4.post0=pypi_0
async-timeout=3.0.1=py_1000
attrs=21.4.0=pyhd8ed1ab_0
brotlipy=0.7.0=pypi_0
bzip2=1.0.8=h1de35cc_0
ca-certificates=2021.10.8=h033912b_0
certifi=2021.10.8=pypi_0
cffi=1.15.0=pypi_0
chardet=4.0.0=pypi_0
charset-normalizer=2.0.10=pyhd8ed1ab_0
click=8.0.3=pypi_0
cryptography=36.0.0=pypi_0
discord.py=1.7.3=pyhd8ed1ab_0
idna=3.3=pyhd8ed1ab_0
libcxx=12.0.0=h2f01273_0
libffi=3.4.2=h0d85af4_5
libzlib=1.2.11=h9173be1_1013
multidict=6.0.2=pypi_0
ncurses=6.2=h2e338ed_4
openssl=3.0.0=h0d85af4_2
pip=21.2.4=pypi_0
pycparser=2.21=pyhd8ed1ab_0
pyopenssl=21.0.0=pyhd8ed1ab_0
pysocks=1.7.1=pypi_0
python=3.10.2=h38b4d05_0_cpython
python-dotenv=0.19.2=pyhd8ed1ab_0
python_abi=3.10=2_cp310
readline=8.1=h05e3726_0
requests=2.27.1=pyhd8ed1ab_0
setuptools=58.0.4=pypi_0
six=1.16.0=pyh6c4a22f_0
sqlite=3.37.0=h707629a_0
tk=8.6.11=h7bc2e8c_0
typing-extensions=4.0.1=hd8ed1ab_0
typing_extensions=4.0.1=pyha770c72_0
tzdata=2021e=hda174b7_0
urllib3=1.26.8=pyhd8ed1ab_1
wheel=0.37.1=pyhd3eb1b0_0
xz=5.2.5=h1de35cc_0
yarl=1.7.2=pypi_0
zlib=1.2.11=h9173be1_1013
